+++ packages/amplify-e2e-tests/src/import-helpers/utilities.ts
@@ -1,4 +1,4 @@
-import { JSONUtilities } from '@aws-amplify/graphql-transformer-core';
+import { JSONUtilities } from '@aws-amplify/amplify-cli-core';
 import {
   addAuthIdentityPoolAndUserPoolWithOAuth,
   addAuthUserPoolOnlyWithOAuth,