+++ packages/amplify-graphql-model-transformer/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@1.3.1...@aws-amplify/graphql-model-transformer@1.3.2) (2023-06-05)
+
+### Bug Fixes
+
+- **graphql:** renamed subscription should generate auth resolver ([bbf3998](https://github.com/aws-amplify/amplify-category-api/commit/bbf399888d65e0e08020da622722d93670bc742c))
+
 ## [1.3.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@1.3.0...@aws-amplify/graphql-model-transformer@1.3.1) (2023-05-23)
 
 **Note:** Version bump only for package @aws-amplify/graphql-model-transformer