+++ packages/graphql-http-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.55](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.54...graphql-http-transformer@5.2.55) (2023-06-05)
+
+**Note:** Version bump only for package graphql-http-transformer
+
 ## [5.2.54](https://github.com/aws-amplify/amplify-category-api/compare/graphql-http-transformer@5.2.53...graphql-http-transformer@5.2.54) (2023-05-23)
 
 **Note:** Version bump only for package graphql-http-transformer